+++ ext_emconf.php
@@ -4,7 +4,7 @@ $EM_CONF[$_EXTKEY] = [
 	'title' => 'Vufind Authentication',
 	'description' => 'Authenticates users based on authenticated vufind session',
 	'category' => 'services',
-	'version' => '3.0.2',
+	'version' => '3.1.0',
 	'state' => 'stable',
 	'author' => 'Ulf Seltmann',
 	'author_email' => 'bdd_dev@ub.uni-leipzig.de',
@@ -12,7 +12,7 @@ $EM_CONF[$_EXTKEY] = [
 	'constraints' => [
 		'depends' => [
 			'php' => '7.4.0-8.1.99',
-			'typo3' => '8.0.0-9.99.99',
+			'typo3' => '9.0.0-9.99.99',
 		],
 		'conflicts' => [
 		],