+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [v3.1.0](https://github.com/ubleipzig/tx-vufind-auth/tree/3.1.0)
+
+[Diff Changelog](https://github.com/ubleipzig/tx-vufind-auth/compare/3.0.2...3.1.0)
+
+* fixes deprecated @inject annotation cmp. [82869](https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/9.0/Feature-82869-ReplaceInjectWithTYPO3CMSExtbaseAnnotationInject.html)
+* removes support for typo3 v8
+* refactoring of cleanup users task to new [Symfony Console Commands](https://docs.typo3.org/m/typo3/reference-coreapi/9.5/en-us/ApiOverview/CommandControllers/Index.html)
+* changes namespace for _AbstractAuthenticationService_ to _\TYPO3\CMS\Core\Authentication_
+* removes _$GLOBAL['TYPO3_DB']_ from _Authentication.php_ [80929](https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/9.0/Breaking-80929-TYPO3_DBMovedToExtension.html)  
+	* replaced by methods at _VufindUserRepository_ 
+
 ## [v3.0.2](https://github.com/ubleipzig/tx-vufind-auth/tree/3.0.2)
 
 [Diff Changelog](https://github.com/ubleipzig/tx-vufind-auth/compare/3.0.1...3.0.2)